refactor: 필터바를 home으로 옮겨 중복되는 코드 줄임 #94
