+++ gomintapa/lib/src/widgets/navigation/home_app_bar.dart
@@ -8,26 +8,24 @@ class HomeAppBar extends StatelessWidget implements PreferredSizeWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBarSection(
-        actions: [
-          IconButton(
-            onPressed: () {
-              // 검색 버튼 클릭 시 동작
-            },
-            icon: Icon(Icons.search_rounded, size: 30),
-          ),
-          SizedBox(width: 10), // 아이콘 버튼 사이 간격
-          IconButton(
-            onPressed: () {
-              // 계정 버튼 클릭 시 MyPage로 이동
-              Navigator.pushNamed(context, '/mypage');
-            },
-            icon: Icon(Icons.account_box_rounded, size: 30),
-          ),
-          SizedBox(width: 10), // 아이콘 버튼 사이 간격
-        ],
-      ),
+    return AppBarSection(
+      actions: [
+        IconButton(
+          onPressed: () {
+            // 검색 버튼 클릭 시 동작
+          },
+          icon: Icon(Icons.search_rounded, size: 30),
+        ),
+        SizedBox(width: 10), // 아이콘 버튼 사이 간격
+        IconButton(
+          onPressed: () {
+            // 계정 버튼 클릭 시 MyPage로 이동
+            Navigator.pushNamed(context, '/mypage');
+          },
+          icon: Icon(Icons.account_box_rounded, size: 30),
+        ),
+        SizedBox(width: 10), // 아이콘 버튼 사이 간격
+      ],
     );
   }
 }
